Little fix, wrong type

+++ RobotArduino/Sensor.cpp
@@ -61,10 +61,11 @@ bool Sensor::bothCloser(int x)
 
 int Sensor::calculateTurnDirection()
 {
-	//If one of the sides distance is lower than 20 decide to not turn to that direction
+	//If one of the sides distance is lower than 20 decide to turn based on distance (what side has the most place to turn)
 	//Otherwise choose random (this is done to escape a square)
-	long centimeterZero = Ultra[0].getCentimeter();
-	long centimeterOne = Ultra[1].getCentimeter();
+
+	int centimeterZero = Ultra[0].getCentimeter();
+	int centimeterOne = Ultra[1].getCentimeter();
 
 	if (centimeterZero < 20 || centimeterOne < 20)
 	{
